fix migration

+++ db/migrate/20140314111316_add_deafault_external_and_internal_role_references_to_department.rb
@@ -1,14 +1,11 @@
 class AddDeafaultExternalAndInternalRoleReferencesToDepartment < ActiveRecord::Migration
 
   def self.up
-    add_column :departments, :default_internal_role_id, :integer
-    add_column :departments, :default_external_role_id, :integer
+    add_column :departments, :default_internal_role_id, :integer unless column_exists?(:departments, :default_internal_role_id)
   end
 
   def self.down
-    remove_column :departments, :default_internal_role_id
-    add_column :departments, :default_external_role_id
-
+    remove_column :departments, :default_internal_role_id if column_exists?(:departments, :default_internal_role_id)
   end
 
 end
